Further prohibit editing of applications when in read only status

+++ applications/views/application_external_locations_views.py
@@ -4,10 +4,11 @@ from rest_framework import status
 from rest_framework.views import APIView
 
 from applications.libraries.case_activity import set_external_location_case_activity
+from applications.libraries.case_status_helpers import get_read_only_case_statuses
 from applications.models import SiteOnApplication, ExternalLocationOnApplication
 from applications.serializers import ExternalLocationOnApplicationSerializer
 from conf.authentication import ExporterAuthentication
-from conf.decorators import authorised_users
+from conf.decorators import authorised_users, application_in_editable_state
 from organisations.libraries.get_external_location import get_location
 from organisations.libraries.get_site import has_previous_sites
 from organisations.models import ExternalLocation
@@ -39,74 +40,79 @@ class ApplicationExternalLocations(APIView):
 
         # Validate that there are actually external locations
         if not location_ids:
-            return JsonResponse(data={'errors': {'external_locations': ['You have to pick at least one external '
-                                                                        'location']
-                                                 }}, status=status.HTTP_400_BAD_REQUEST)
+            return JsonResponse(data={
+                'errors': {'external_locations': ['You have to pick at least one external location']}
+                }, status=status.HTTP_400_BAD_REQUEST)
 
         previous_locations = ExternalLocationOnApplication.objects.filter(application=application)
         previous_location_ids = [str(previous_location_id) for previous_location_id in
                                  previous_locations.values_list('external_location__id', flat=True)]
 
-        if not application.status or application.status.status == CaseStatusEnum.APPLICANT_EDITING:
-            new_locations = [get_location(location_id, request.user.organisation) for location_id in location_ids if
-                             location_id not in previous_location_ids]
-        else:
-            if has_previous_sites(application):
-                return JsonResponse(data={'errors': {
+        if application.status and application.status.status in get_read_only_case_statuses():
+            return JsonResponse(data={'errors': {
                     'external_locations': [
-                        "Go back and change your answer from ‘Change a site, or delete a good, third party or "
-                        "country’ to ’Change something else’."]
+                        f'Application status {application.status.status} is read-only.']
                 }}, status=status.HTTP_400_BAD_REQUEST)
-
-            previous_location_countries = list(previous_locations.values_list('external_location__country__id',
-                                                                              flat=True))
-            new_locations = []
-
-            for location_id in location_ids:
-                new_location = get_location(location_id, request.user.organisation)
-
-                if new_location.country.id not in previous_location_countries:
+        else:
+            if not application.status or application.status.status == CaseStatusEnum.APPLICANT_EDITING:
+                new_locations = [get_location(location_id, request.user.organisation)
+                                 for location_id in location_ids if location_id not in previous_location_ids]
+            else:
+                if has_previous_sites(application):
                     return JsonResponse(data={'errors': {
                         'external_locations': [
                             "Go back and change your answer from ‘Change a site, or delete a good, third party or "
                             "country’ to ’Change something else’."]
                     }}, status=status.HTTP_400_BAD_REQUEST)
-                elif str(new_location.id) not in previous_location_ids:
-                    new_locations.append(new_location)
 
-        # Update draft activity
-        application.activity = 'Brokering'
-        application.save()
+                previous_location_countries = list(previous_locations.values_list('external_location__country__id',
+                                                                              flat=True))
+                new_locations = []
 
-        # Get locations on to be removed
-        removed_locations = []
-        if data.get('method') != 'append_location':
-            removed_location_ids = list(set(previous_location_ids) - set(location_ids))
-            removed_locations = previous_locations.filter(external_location__id__in=removed_location_ids)
+                for location_id in location_ids:
+                    new_location = get_location(location_id, request.user.organisation)
+                    if previous_location_countries and new_location.country.id not in previous_location_countries:
+                        return JsonResponse(data={'errors': {
+                            'external_locations': [
+                                "Go back and change your answer from ‘Change a site, or delete a good, third party or "
+                                "country’ to ’Change something else’."]
+                        }}, status=status.HTTP_400_BAD_REQUEST)
+                    elif str(new_location.id) not in previous_location_ids:
+                        new_locations.append(new_location)
 
-        # Append new ExternalLocationOnApplications
-        response_data = []
-        for new_location in new_locations:
-            serializer = ExternalLocationOnApplicationSerializer(
-                data={'external_location': new_location.id, 'application': application.id})
+            # Update activity
+            application.activity = 'Brokering'
+            application.save()
 
-            if serializer.is_valid():
-                serializer.save()
-                response_data.append(serializer.data)
-            else:
-                return JsonResponse(data={'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+            # Get locations to be removed
+            removed_locations = []
+            if data.get('method') != 'append_location':
+                removed_location_ids = list(set(previous_location_ids) - set(location_ids))
+                removed_locations = previous_locations.filter(external_location__id__in=removed_location_ids)
 
-        # Get sites to be removed if a site is being added
-        removed_sites = SiteOnApplication.objects.filter(application=application)
+            # Append new ExternalLocationOnApplications
+            response_data = []
+            for new_location in new_locations:
+                serializer = ExternalLocationOnApplicationSerializer(
+                    data={'external_location': new_location.id, 'application': application.id})
 
-        set_external_location_case_activity(removed_sites, removed_locations, new_locations, request.user, application)
+                if serializer.is_valid():
+                    serializer.save()
+                    response_data.append(serializer.data)
+                else:
+                    return JsonResponse(data={'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
-        if data.get('method') != 'append_location':
-            removed_locations.delete()
+            # Get sites to be removed if a site is being added
+            removed_sites = SiteOnApplication.objects.filter(application=application)
 
-        removed_sites.delete()
+            set_external_location_case_activity(removed_sites, removed_locations, new_locations, request.user, application)
 
-        return JsonResponse(data={'external_locations': response_data}, status=status.HTTP_201_CREATED)
+            if data.get('method') != 'append_location':
+                removed_locations.delete()
+
+            removed_sites.delete()
+
+            return JsonResponse(data={'external_locations': response_data}, status=status.HTTP_201_CREATED)
 
 
 class ApplicationRemoveExternalLocation(APIView):
@@ -114,17 +120,20 @@ class ApplicationRemoveExternalLocation(APIView):
 
     @authorised_users(ExporterUser)
     def delete(self, request, application, ext_loc_pk):
-        if application.status and application.status.status != CaseStatusEnum.APPLICANT_EDITING:
-            if ExternalLocationOnApplication.objects.filter(application=application).count() == 1:
-                return JsonResponse(data={'error': "Go back and change your answer from ‘Change a site, or delete a "
-                                                     "good, third party or country’ to ’Change something else’."},
-                                    status=status.HTTP_400_BAD_REQUEST)
+        if application.status and application.status.status in get_read_only_case_statuses():
+            return JsonResponse(data={'error': f'Application status {application.status.status} is read-only.'},
+                                status=status.HTTP_400_BAD_REQUEST)
+        else:
+            if application.status and application.status.status != CaseStatusEnum.APPLICANT_EDITING:
+                if ExternalLocationOnApplication.objects.filter(application=application).count() == 1:
+                    return JsonResponse(data={'error': "Go back and change your answer from ‘Change a site, or delete "
+                                                       "a good, third party or country’ to ’Change something else’."},
+                                        status=status.HTTP_400_BAD_REQUEST)
 
-        removed_locations = ExternalLocationOnApplication.objects.filter(application=application,
-                                                                         external_location__pk=ext_loc_pk)
+            removed_locations = ExternalLocationOnApplication.objects.filter(application=application,
+                                                                             external_location__pk=ext_loc_pk)
 
-        set_external_location_case_activity(None, removed_locations, None, request.user, application)
+            set_external_location_case_activity(None, removed_locations, None, request.user, application)
+            removed_locations.delete()
 
-        removed_locations.delete()
-
-        return JsonResponse(data={'success': 'External location deleted'}, status=status.HTTP_204_NO_CONTENT)
+            return JsonResponse(data={'success': 'External location deleted'}, status=status.HTTP_204_NO_CONTENT)
